(oops! sorry!) Fix warnings introduced by previous commit

+++ cpu.c
@@ -2203,12 +2203,12 @@ void ARCH_DEP( process_trace )( REGS* regs, BYTE* dest )
     {
         if (sysblk.breakasid)
         {
-            if (sysblk.breakasid_arn == USE_PRIMARY_SPACE)
+            if (sysblk.breakasid_arn == (BYTE)USE_PRIMARY_SPACE)
             {
                 if (regs->CR_LHL(4) != sysblk.breakasid)
                     regs->stepping = false;
             }
-            else if (sysblk.breakasid_arn == USE_HOME_SPACE)
+            else if (sysblk.breakasid_arn == (BYTE)USE_HOME_SPACE)
             {
                 if (regs->CR_LHL(3) != sysblk.breakasid)
                     regs->stepping = false;
